src/requirements.txt
celery==5.4.0
django==5.1.4
django-celery-beat==2.7.0
djangorestframework==3.14.0
django_jazzmin==3.0.1
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
python-dotenv==1.0.1
isort==5.12.0
psycopg2-binary==2.9.9
selenium==4.27.1
setuptools==75.6.0
redis==5.2.1
requests==2.32.3